Eamples: Added support for --range and --log command line arguments in the s500, omniscan450, and surveyor240 examples.
